+++ tests/urls.py
@@ -15,6 +15,7 @@ urlpatterns = [
     path("async/streaming/", views.async_streaming),
     path("async/streaming/empty/", views.async_streaming_empty),
     path("async/streaming/blanks/", views.async_streaming_blanks),
+    path("random/", views.random),
     path("binary/", views.binary),
     path("etag/", views.etag),
 ]